史诗：QQ 端删除及展示错题

+++ internal/app/datahub/public/cuoti/struct.go
@@ -2,7 +2,7 @@ package cuoti
 
 type Tab struct {
 	ID             uint32 `json:"id" db:"id"`                             // 唯一标识符
-	UserID         uint32 `json:"user_id" db:"user_id"`                   // 用户 QQ 号
+	UserID         uint64 `json:"user_id" db:"user_id"`                   // 用户 QQ 号
 	Question       string `json:"question" db:"question"`                 // 错题题目
 	ImageURL       string `json:"image_url" db:"image_url"`               // 错题图片 URL
 	QuestionAnswer string `json:"question_answer" db:"question_answer"`   // 问题正确答案
